+++ apps/modernization-ui/src/apps/deduplication/api/model/AlgorithmExport.ts
@@ -1,0 +1,7 @@
+import { DataElements } from 'apps/deduplication/api/model/DataElement';
+import { Pass } from './Pass';
+
+export type AlgorithmExport = {
+    dataElements: DataElements;
+    algorithm: { passes: Pass[] };
+};
